medication_sku_catalogue/requirements.txt
Django
djangorestframework
django-filter
djangorestframework-simplejwt
drf_spectacular
drf_yasg